+++ src/test/java/com/linuxforhealth/connect/processor/BlueButton20ResultTest.java
@@ -6,7 +6,7 @@
 package com.linuxforhealth.connect.processor;
 
 import ca.uhn.fhir.context.FhirContext;
-import com.linuxforhealth.connect.TestUtils;
+import com.linuxforhealth.connect.support.TestUtils;
 import org.apache.camel.Exchange;
 import org.apache.camel.ExtendedExchange;
 import org.apache.camel.support.DefaultExchange;
