+++ src/test/java/com/linuxforhealth/connect/support/TestUtils.java
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.linuxforhealth.connect;
+package com.linuxforhealth.connect.support;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
